extract and improve advance
