+++ src/trace/layers/ordered.rs
@@ -1,6 +1,6 @@
 //! Implementation using ordered keys and exponential search.
 
-use super::{Trie, Cursor, Builder, MergeBuilder, TupleBuilder};
+use super::{Trie, Cursor, Builder, MergeBuilder, TupleBuilder, advance};
 
 /// A level of the trie, with keys and offsets into a lower layer.
 ///
@@ -36,19 +36,19 @@ impl<K: Ord+Clone, L: Trie> Trie for OrderedLayer<K, L> {
 				bounds: (lower, upper),
 				child: self.vals.cursor_from(child_lower, child_upper),
 				pos: lower,
-			}		
+			}
 		}
 		else {
 			OrderedCursor {
 				bounds: (0, 0),
 				child: self.vals.cursor_from(0, 0),
 				pos: 0,
-			}	
+			}
 		}
 	}
 }
 
-/// Assembles a layer of this 
+/// Assembles a layer of this
 pub struct OrderedBuilder<K: Ord, L> {
 	/// Keys
 	pub keys: Vec<K>,
@@ -59,8 +59,8 @@ pub struct OrderedBuilder<K: Ord, L> {
 }
 
 impl<K: Ord+Clone, L: Builder> Builder for OrderedBuilder<K, L> {
-	type Trie = OrderedLayer<K, L::Trie>; 
-	fn boundary(&mut self) -> usize { 
+	type Trie = OrderedLayer<K, L::Trie>;
+	fn boundary(&mut self) -> usize {
 		self.offs[self.keys.len()] = self.vals.boundary();
 		self.keys.len()
 	}
@@ -162,12 +162,12 @@ impl<K: Ord+Clone, L: TupleBuilder> TupleBuilder for OrderedBuilder<K, L> {
 
 	type Item = (K, L::Item);
 	fn new() -> Self { OrderedBuilder { keys: Vec::new(), offs: vec![0], vals: L::new() } }
-	fn with_capacity(cap: usize) -> Self { 
+	fn with_capacity(cap: usize) -> Self {
 		let mut offs = Vec::with_capacity(cap + 1);
 		offs.push(0);
-		OrderedBuilder{ 
-			keys: Vec::with_capacity(cap), 
-			offs: offs, 
+		OrderedBuilder{
+			keys: Vec::with_capacity(cap),
+			offs: offs,
 			vals: L::with_capacity(cap),
 		}
 	}
@@ -201,7 +201,7 @@ impl<K: Ord, L: Trie> Cursor<OrderedLayer<K, L>> for OrderedCursor<L> {
 	type Key = K;
 	fn key<'a>(&self, storage: &'a OrderedLayer<K, L>) -> &'a Self::Key { &storage.keys[self.pos] }
 	fn step(&mut self, storage: &OrderedLayer<K, L>) {
-		self.pos += 1; 
+		self.pos += 1;
 		if self.valid(storage) {
 			self.child.reposition(&storage.vals, storage.offs[self.pos], storage.offs[self.pos + 1]);
 		}
@@ -230,39 +230,4 @@ impl<K: Ord, L: Trie> Cursor<OrderedLayer<K, L>> for OrderedCursor<L> {
 			self.child.reposition(&storage.vals, storage.offs[self.pos], storage.offs[self.pos + 1]);
 		}
 	}
-}
-
-/// Reports the number of elements satisfing the predicate.
-///
-/// This methods *relies strongly* on the assumption that the predicate
-/// stays false once it becomes false, a joint property of the predicate
-/// and the slice. This allows `advance` to use exponential search to 
-/// count the number of elements in time logarithmic in the result.
-#[inline(never)]
-pub fn advance<T, F: Fn(&T)->bool>(slice: &[T], function: F) -> usize {
-
-	// start with no advance
-	let mut index = 0;
-	if index < slice.len() && function(&slice[index]) {
-
-		// advance in exponentially growing steps.
-		let mut step = 1;
-		while index + step < slice.len() && function(&slice[index + step]) {
-			index += step;
-			step = step << 1;
-		}
-
-		// advance in exponentially shrinking steps.
-		step = step >> 1;
-		while step > 0 {
-			if index + step < slice.len() && function(&slice[index + step]) {
-				index += step;
-			}
-			step = step >> 1;
-		}
-
-		index += 1;
-	}	
-
-	index
 }
